enrichment-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.12
imageHash==4.3.2
pillow==11.2.1
pydantic==2.11.3
python-dotenv==1.1.0
requests==2.32.3
uvicorn==0.34.1
boto3==1.37.31
torch==2.6.0
torchvision==0.21.0
psycopg2-binary==2.9.10
python-multipart==0.0.20